- Removed some functions in AccessControlManager due to already refined and migrated to new implementation - Added new action result statuses - Updated base repository functions - Added new coroutine function to return deferred result - Updated getTabCompleteSuggestion function - Refined CommandManager - Refined GuiManager - Refined 'unassignPlayerRole' api function - Refined 'getRole' api function - Refined 'addRolePermissionList' api function - Separated class for PermissionCommandDispatcher - Refined PermissionManager - Commented some old api function - Refined PlayerSessionManagerNew - Separated class for RoleCommandDispatcher - Removed some functions in RoleManager(old) due to already refined and migrated to new implementation - Refined RoleManagerNew

+++ neon-kotlin/neon-api-new/src/main/kotlin/com/islandstudio/neon/apinew/facade/player/IRolePermissionFacade.kt
@@ -5,7 +5,7 @@ import com.islandstudio.neon.apinew.dto.action.player.permission.AddRolePermisio
 import com.islandstudio.neon.apinew.dto.result.player.permission.RolePermissionDetailResultDTO
 
 interface IRolePermissionFacade {
-    suspend fun addRolePermissionList(action: AddRolePermisionActionDTO): ResultProvider<Int>
+    suspend fun addRolePermissionList(action: AddRolePermisionActionDTO): ResultProvider<Long>
     suspend fun getRolePermissionDetailList(roleId: Long): ResultProvider<ArrayList<RolePermissionDetailResultDTO>>
     suspend fun removeRolePermissionListByIds(ids: ArrayList<Long>): ResultProvider<Boolean>
     suspend fun isRolePermissionExistByPermissionCodes(roleCode: String, permissionCodes: ArrayList<String>): ResultProvider<Boolean>
